requirements.txt
amqp==2.3.2
anyjson==0.3.3
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
ddt==1.2.0
Django==2.1.3
django-celery==3.1.0
django-crispy-forms==1.7.2
django-suit==0.2.26
django-utils2==2.5.0
django-waffle==0.15.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
factory-boy==2.11.1
Faker==1.0.0
idna==2.8
kombu==4.2.2
mock==2.0.0
numpy==1.15.4
pbr==5.1.1
phonenumbers==8.10.0
Pillow==5.3.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
PyJWT==1.6.4
python-dateutil==2.7.5
python-utils==2.3.0
pytz==2018.7
requests==2.21.0
six==1.11.0
text-unidecode==1.2
urllib3==1.24.1
vine==1.1.4